Add a Japanese segmenter
